principal 커스텀했기에 사용법 변경필요. 로그인시 user ID 값도 내려주게 하기. rating Float 로 변경.

+++ core/src/main/kotlin/app/foodin/domain/review/Review.kt
@@ -34,7 +34,7 @@ data class Review(
 
     var commentCount: Int = 0
 
-    var rating: Int = 0
+    var rating: Float = 0F
 
     constructor(reviewCreateReq: ReviewCreateReq): this(foodId = reviewCreateReq.foodId) {
         reviewCreateReq.let {
